fix(developer): handle errors parsing .kps file when loading project
When a project is loaded, the IDE parses .kps files to retrieve metadata
for viewing. If the .kps file was malformed, then the parser could fail,
blocking the load of the project. Instead, we just ignore the errors and
show no metadata. Note that the Package Editor already loads the file
successfully and reports on the problem, allowing the user to fix it.

Fixes: #11992
Fixes: KEYMAN-DEVELOPER-23W