chore(web): remove `paths` from `tsconfig.json`
See #12027 (as well as build failures in the current repo restructurings)
for the motivation for this change.

Basically remove all `path` lines from `references` except when they
reference files in or below the current directory. Another execption is
if there is no `package.json` file for the current directory, i.e. the
current directory is not a separate module.