refactor(developer): Move error reporting inside `CheckForDuplicateStore`
This is the same change as for `CheckForDuplicateGroup`.

`CheckForDuplicateStore` is one of a few places where parameterized
errors are returned. Moving the error reporting inside the function
makes it cleaner to refactor those parameterized errors later. (This
requires a similar change in `ProcessStoreLine`, also formatted and
rebraced per style guide.)