fix(core): fix generated `core.js` file for debug builds
Due to emscripten bug emscripten-core/emscripten#22754 specifying only
`webview` as environment resulted in a broken `core.js` file when doing
a debug build. In release builds things worked because of the
`-Wl,-O1 -O2` that meson added for the release build. This change works
around that bug by adding `web` as additional environment.