feat(developer): use Windows `isAbsolute` to avoid cross-platform mismatches
As our data files often come from Windows, we need to test if paths are
Windows-style absolute paths, with either path separator (e.g. `C:\...`,
`\path\...`, `\\server\path\...`, `/path/...`). Technically, this means
a posix path that starts with an escaped character would be regarded as
absolute, but IMHO those kinds of shenanigans deserve to result in a
failing build anyway. ;-)