fix(windows): check params status before download
In the case when a Keyboard was available for update but the
keyman version wasn't the state machine correctly didn't try to install
it. However the back ground download loop would try to download and
empty Params.InstallURL. This commit now checks the Params.Status.
This commit also removes the dead code for TDownloadUpdateParams record
type. This code is left over from a progress ui callback. It was
initially left for when UI callback maybe added back in. However, it
is currently confusing deadcode so has been removed.
Fixes: #13153