feat(windows): hack in some fun features for kmdevlink
This was an evening activity scratching itches on kmdevlink's 'Open
Issue' dialog. The search box now supports three different types of
strings:

1. `[[repo]#]num` -> open issue/PR #num in repo (if repo omitted, will
   use 'keyman'):
   * `13235` or `#13235` or `keyman#13235` -> opens PR #13235
   * `keyman.com#545` -> opens PR keymanapp/keyman.com#545
   * `api#172` -> opens PR keymanapp/api.keyman.com#172
2. `[[repo]#]searchtext` -> any non-numeric string does an issue search:
   * `label:ios/ font` -> opens an issue search for 'ios/' label and
     'font' in keymanapp/keyman repo
   * `api#db` -> opens an issue search for 'db' in
     keymanapp/api.keyman.com repo
3. `[[repo]#]` -> opens issue list for repo (keymanapp/keyman if repo is
   omitted, or empty box)

A second feature is construction of a HTML snippet of our usual
short-form links to issues/PRs for Google Docs. Note that this dialog
does not check if the target is an issue or PR, it just builds a link to
the issue URL and lets GitHub redirect to PR. To use this, just type the
`[[repo]#]num` format into the search field and click the Copy button.