fix(windows): automaticaly report usage defualt false in options
The process is quite confusing. The options from the setup install
options will only resualt in a command line option being past through
to firstrun if it is true. If is false the options will not be passed
through. Therefore if this is the first installation once when the
Keyman Configuration is opened it will show the default option
configured in utilkeymanoptions.pas, in kmcomapi. This result of all
this was a user would deselect the option in setup. When they opened
configuration it was set to true.