chore(linux): add TC messages for integration tests
This allows TeamCity to see and report the individual tests. Unfortunately
GLib's gtest doesn't have support for TC and doesn't provide hooks that
we could use, so this change copied the gtest files and renamed the
methods to `km_g_test_*` and added the necessary output.

Also add a `--teamcity` parameter to `run-tests.sh`.