fix(windows): sets options values in firstrun
Previously it only set the startwithwindows, checkforupdats and
autmaticallyreportusage values in firstrun if the flag was true.
It was only true if the argument was passed to the install executable on
the command line. However, it is valid to treat its non existance on the
command line as false for these 3 values.