fix(linux): improve emitting keystrokes
For some reason when using a LDML keyboard we ended up emitting a
keystroke even though Core returned `FALSE for `emitting_keystroke`.
I wasn't able to track down where this happened, but this change seems
like an easy way to fix it.

This showed up when debugging why the changes in #13372 didn't work for
non-compliant apps.