feat(web): bundles and minifies the new trie-decoding function
By bundling and minifying the trie-decoding function properly, we gain a build output that can be included with compiled lexical models to maintain backward-compatibility in older versions of the predictive-text engine.

This commit only covers the Web side; this code is not yet set for inclusion within compiled lexical models.