fix(ios): prevent re-display of keyboard when re-entering app to its menus
Fixes: #13629
Fixes: #11953
Fixes: #6387

This change makes sure that upon app re-entry, the keyboard only displays when the main view will be the active view.  If the app had been in a menu or in the picker, the keyboard will no longer show up on top of it.  Additionally, if the app is re-entered to directly install a KMP from a file, the keyboard will not overlay the package installer.

The core of the problem is that upon app re-entry, the WebView reloads the keyboard, which in turns triggers the `keyboardLoaded` internal notification.  That notification did not check if the main view was the currently-visible view or not, and would always display the keyboard.