fix(ios): do not notify keyboard download when showing package installer for it
Fixes: #12590

If we're already showing the package-installer for a keyboard the user found via search... then we clearly downloaded it successfully.  Why double-notify?

It turns out that #12590 itself was due to our usage of a TabBarController outside the prescribed safe use cases.  We're actually fine most of the time... but it seems the tab area doesn't play nicely with a NavigationController toolbar, which is what displays download + update notifications.  Rather than find a way to force that to work, it's simpler to just... not show the toolbar, **ever**, when the package installer is displayed.