maint(common): consolidate VC++ environment setup for meson
Use the automated environment setup that we built for Windows and
Developer builds in our meson build configurations, so we don't need to
call out to vsdevcmd.bat for every build step. This makes the build
faster, more consistent across the various ways we call VC++, and
simplifies the shell scripts. Removes the complicated build.bat wrappers.

Also removes the `cleanup_visual_studio_path` function which should no
longer be necessary, as our environment builder does the job for us.

Minor complication in that `$TARGET_PATH` variable in kmcmplib/build.sh
is also used in the Windows environment setup, so eliminated its use in
kmcmplib/build.sh.

Finally, removes the `$MESON_LOW_VERSION` variable as we know we are
always on Meson 1.0 or later now.

Fixes: #12639