chore(web): add accessor annotations for `JSKeyboardProcessor`
Preparation for epic/web-core, clarify which members of
`JSKeyboardProcessor` are private and which will need equivalent
implementations in `CoreKeyboardProcessor`.

Note: some of the publicly accessed members are not immediately obvious,
because they are accessed via interfaces declared in other modules (for
example `baseLayout` is referenced in `KeyboardState` type in
hardwareEventKeyboard.ts).